split out tests for easier editing
